requirements.txt
beautifulsoup4
requests
pytz
discord-webhook
python-dotenv